pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medtriage_env"
version = "1.0.0"
description = "Clinical triage and diagnostic reasoning environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core==0.2.3",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "openai>=2.7.2",
    "python-dotenv",
    "httpx>=0.28.1",
]

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
